Rn074 example (#3408)
* chore: upgrade example to rn-0.74.0-rc.2

* rn 0.74 ios updates

* fix: fix kotlin corutines to 1.8.0 instead of binding to kotlin version

+++ android/src/main/old-arch/com/rnmapbox/rnmbx/NativeRNMBXPointAnnotationModuleSpec.java
@@ -36,5 +36,5 @@ public abstract class NativeRNMBXPointAnnotationModuleSpec extends ReactContextB
 
   @ReactMethod
   @DoNotStrip
-  public abstract void refresh(@Nullable Double viewRef, Promise promise);
+  public abstract void refresh(@Nullable Integer viewRef, Promise promise);
 }
